refacto manager + separate test access
