+++ public/test-coverage/Controller/dashboard.html
@@ -59,6 +59,7 @@
        </thead>
        <tbody>
        <tr><td><a href="SecurityController.php.html#11">App\Controller\SecurityController</a></td><td class="text-right">85%</td></tr>
+       <tr><td><a href="UserController.php.html#21">App\Controller\UserController</a></td><td class="text-right">88%</td></tr>
 
        </tbody>
       </table>
@@ -75,6 +76,7 @@
         </tr>
        </thead>
        <tbody>
+       <tr><td><a href="UserController.php.html#21">App\Controller\UserController</a></td><td class="text-right">9</td></tr>
 
        </tbody>
       </table>
@@ -113,7 +115,8 @@
        </thead>
        <tbody>
        <tr><td><a href="SecurityController.php.html#31"><abbr title="App\Controller\SecurityController::logout">logout</abbr></a></td><td class="text-right">0%</td></tr>
-       <tr><td><a href="TaskController.php.html#33"><abbr title="App\Controller\TaskController::listAction">listAction</abbr></a></td><td class="text-right">88%</td></tr>
+       <tr><td><a href="UserController.php.html#96"><abbr title="App\Controller\UserController::deleteTaskAction">deleteTaskAction</abbr></a></td><td class="text-right">0%</td></tr>
+       <tr><td><a href="TaskController.php.html#40"><abbr title="App\Controller\TaskController::listAction">listAction</abbr></a></td><td class="text-right">88%</td></tr>
 
        </tbody>
       </table>
@@ -130,7 +133,7 @@
         </tr>
        </thead>
        <tbody>
-       <tr><td><a href="TaskController.php.html#33"><abbr title="App\Controller\TaskController::listAction">listAction</abbr></a></td><td class="text-right">3</td></tr>
+       <tr><td><a href="TaskController.php.html#40"><abbr title="App\Controller\TaskController::listAction">listAction</abbr></a></td><td class="text-right">3</td></tr>
 
        </tbody>
       </table>
@@ -140,7 +143,7 @@
    <footer>
     <hr/>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
@@ -159,7 +162,7 @@ $(document).ready(function() {
       .yAxis.tickFormat(d3.format('d'));
 
     d3.select('#classCoverageDistribution svg')
-      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,1,1,2], "Class Coverage"))
+      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,2,1,1], "Class Coverage"))
       .transition().duration(500).call(chart);
 
     nv.utils.windowResize(chart.update);
@@ -177,7 +180,7 @@ $(document).ready(function() {
       .yAxis.tickFormat(d3.format('d'));
 
     d3.select('#methodCoverageDistribution svg')
-      .datum(getCoverageDistributionData([1,0,0,0,0,0,0,0,0,1,0,12], "Method Coverage"))
+      .datum(getCoverageDistributionData([2,0,0,0,0,0,0,0,0,1,0,11], "Method Coverage"))
       .transition().duration(500).call(chart);
 
     nv.utils.windowResize(chart.update);
@@ -227,7 +230,7 @@ $(document).ready(function() {
     chart.yAxis.axisLabel('Cyclomatic Complexity');
 
     d3.select('#classComplexity svg')
-      .datum(getComplexityData([[100,1,"<a href=\"HomePageController.php.html#10\">App\\Controller\\HomePageController<\/a>"],[85.71428571428571,2,"<a href=\"SecurityController.php.html#11\">App\\Controller\\SecurityController<\/a>"],[98,13,"<a href=\"TaskController.php.html#21\">App\\Controller\\TaskController<\/a>"],[100,9,"<a href=\"UserController.php.html#20\">App\\Controller\\UserController<\/a>"]], 'Class Complexity'))
+      .datum(getComplexityData([[100,1,"<a href=\"HomePageController.php.html#10\">App\\Controller\\HomePageController<\/a>"],[85.71428571428571,2,"<a href=\"SecurityController.php.html#11\">App\\Controller\\SecurityController<\/a>"],[98.0392156862745,13,"<a href=\"TaskController.php.html#22\">App\\Controller\\TaskController<\/a>"],[88,9,"<a href=\"UserController.php.html#21\">App\\Controller\\UserController<\/a>"]], 'Class Complexity'))
       .transition()
       .duration(500)
       .call(chart);
@@ -251,7 +254,7 @@ $(document).ready(function() {
     chart.yAxis.axisLabel('Method Complexity');
 
     d3.select('#methodComplexity svg')
-      .datum(getComplexityData([[100,1,"<a href=\"HomePageController.php.html#15\">App\\Controller\\HomePageController::index<\/a>"],[100,1,"<a href=\"SecurityController.php.html#16\">App\\Controller\\SecurityController::login<\/a>"],[0,1,"<a href=\"SecurityController.php.html#31\">App\\Controller\\SecurityController::logout<\/a>"],[100,1,"<a href=\"TaskController.php.html#24\">App\\Controller\\TaskController::__construct<\/a>"],[88.88888888888889,3,"<a href=\"TaskController.php.html#33\">App\\Controller\\TaskController::listAction<\/a>"],[100,3,"<a href=\"TaskController.php.html#54\">App\\Controller\\TaskController::createAction<\/a>"],[100,3,"<a href=\"TaskController.php.html#80\">App\\Controller\\TaskController::editAction<\/a>"],[100,1,"<a href=\"TaskController.php.html#107\">App\\Controller\\TaskController::deleteTaskAction<\/a>"],[100,2,"<a href=\"TaskController.php.html#122\">App\\Controller\\TaskController::toggleTaskAction<\/a>"],[100,1,"<a href=\"UserController.php.html#23\">App\\Controller\\UserController::__construct<\/a>"],[100,1,"<a href=\"UserController.php.html#33\">App\\Controller\\UserController::getUserList<\/a>"],[100,3,"<a href=\"UserController.php.html#46\">App\\Controller\\UserController::createAction<\/a>"],[100,3,"<a href=\"UserController.php.html#69\">App\\Controller\\UserController::editAction<\/a>"],[100,1,"<a href=\"UserController.php.html#90\">App\\Controller\\UserController::deleteTaskAction<\/a>"]], 'Method Complexity'))
+      .datum(getComplexityData([[100,1,"<a href=\"HomePageController.php.html#15\">App\\Controller\\HomePageController::index<\/a>"],[100,1,"<a href=\"SecurityController.php.html#16\">App\\Controller\\SecurityController::login<\/a>"],[0,1,"<a href=\"SecurityController.php.html#31\">App\\Controller\\SecurityController::logout<\/a>"],[100,1,"<a href=\"TaskController.php.html#30\">App\\Controller\\TaskController::__construct<\/a>"],[88.88888888888889,3,"<a href=\"TaskController.php.html#40\">App\\Controller\\TaskController::listAction<\/a>"],[100,3,"<a href=\"TaskController.php.html#61\">App\\Controller\\TaskController::createAction<\/a>"],[100,3,"<a href=\"TaskController.php.html#87\">App\\Controller\\TaskController::editAction<\/a>"],[100,1,"<a href=\"TaskController.php.html#114\">App\\Controller\\TaskController::deleteTaskAction<\/a>"],[100,2,"<a href=\"TaskController.php.html#129\">App\\Controller\\TaskController::toggleTaskAction<\/a>"],[100,1,"<a href=\"UserController.php.html#29\">App\\Controller\\UserController::__construct<\/a>"],[100,1,"<a href=\"UserController.php.html#39\">App\\Controller\\UserController::getUserList<\/a>"],[100,3,"<a href=\"UserController.php.html#52\">App\\Controller\\UserController::createAction<\/a>"],[100,3,"<a href=\"UserController.php.html#75\">App\\Controller\\UserController::editAction<\/a>"],[0,1,"<a href=\"UserController.php.html#96\">App\\Controller\\UserController::deleteTaskAction<\/a>"]], 'Method Complexity'))
       .transition()
       .duration(500)
       .call(chart);
