+++ public/test-coverage/Controller/index.html
@@ -45,29 +45,29 @@
       <tr>
        <td class="success">Total</td>
        <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="97.56" aria-valuemin="0" aria-valuemax="100" style="width: 97.56%">
-           <span class="sr-only">97.56% covered (success)</span>
+         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="94.05" aria-valuemin="0" aria-valuemax="100" style="width: 94.05%">
+           <span class="sr-only">94.05% covered (success)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">97.56%</div></td>
-       <td class="success small"><div align="right">80&nbsp;/&nbsp;82</div></td>
+       <td class="success small"><div align="right">94.05%</div></td>
+       <td class="success small"><div align="right">79&nbsp;/&nbsp;84</div></td>
        <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="85.71" aria-valuemin="0" aria-valuemax="100" style="width: 85.71%">
-           <span class="sr-only">85.71% covered (warning)</span>
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="78.57" aria-valuemin="0" aria-valuemax="100" style="width: 78.57%">
+           <span class="sr-only">78.57% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="warning small"><div align="right">85.71%</div></td>
-       <td class="warning small"><div align="right">12&nbsp;/&nbsp;14</div></td>
+       <td class="warning small"><div align="right">78.57%</div></td>
+       <td class="warning small"><div align="right">11&nbsp;/&nbsp;14</div></td>
        <td class="danger big">       <div class="progress">
-         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="50.00" aria-valuemin="0" aria-valuemax="100" style="width: 50.00%">
-           <span class="sr-only">50.00% covered (danger)</span>
+         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="25.00" aria-valuemin="0" aria-valuemax="100" style="width: 25.00%">
+           <span class="sr-only">25.00% covered (danger)</span>
          </div>
        </div>
 </td>
-       <td class="danger small"><div align="right">50.00%</div></td>
-       <td class="danger small"><div align="right">2&nbsp;/&nbsp;4</div></td>
+       <td class="danger small"><div align="right">25.00%</div></td>
+       <td class="danger small"><div align="right">1&nbsp;/&nbsp;4</div></td>
       </tr>
 
       <tr>
@@ -129,13 +129,13 @@
       <tr>
        <td class="success"><img src="../_icons/file-code.svg" class="octicon" /><a href="TaskController.php.html">TaskController.php</a></td>
        <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="98.00" aria-valuemin="0" aria-valuemax="100" style="width: 98.00%">
-           <span class="sr-only">98.00% covered (success)</span>
+         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="98.04" aria-valuemin="0" aria-valuemax="100" style="width: 98.04%">
+           <span class="sr-only">98.04% covered (success)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">98.00%</div></td>
-       <td class="success small"><div align="right">49&nbsp;/&nbsp;50</div></td>
+       <td class="success small"><div align="right">98.04%</div></td>
+       <td class="success small"><div align="right">50&nbsp;/&nbsp;51</div></td>
        <td class="warning big">       <div class="progress">
          <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="83.33" aria-valuemin="0" aria-valuemax="100" style="width: 83.33%">
            <span class="sr-only">83.33% covered (warning)</span>
@@ -155,31 +155,31 @@
       </tr>
 
       <tr>
-       <td class="success"><img src="../_icons/file-code.svg" class="octicon" /><a href="UserController.php.html">UserController.php</a></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+       <td class="warning"><img src="../_icons/file-code.svg" class="octicon" /><a href="UserController.php.html">UserController.php</a></td>
+       <td class="warning big">       <div class="progress">
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="88.00" aria-valuemin="0" aria-valuemax="100" style="width: 88.00%">
+           <span class="sr-only">88.00% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">24&nbsp;/&nbsp;24</div></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+       <td class="warning small"><div align="right">88.00%</div></td>
+       <td class="warning small"><div align="right">22&nbsp;/&nbsp;25</div></td>
+       <td class="warning big">       <div class="progress">
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="80.00" aria-valuemin="0" aria-valuemax="100" style="width: 80.00%">
+           <span class="sr-only">80.00% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">5&nbsp;/&nbsp;5</div></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+       <td class="warning small"><div align="right">80.00%</div></td>
+       <td class="warning small"><div align="right">4&nbsp;/&nbsp;5</div></td>
+       <td class="danger big">       <div class="progress">
+         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
+           <span class="sr-only">0.00% covered (danger)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">1&nbsp;/&nbsp;1</div></td>
+       <td class="danger small"><div align="right">0.00%</div></td>
+       <td class="danger small"><div align="right">0&nbsp;/&nbsp;1</div></td>
       </tr>
 
 
@@ -195,7 +195,7 @@
      <span class="success"><strong>High</strong>: 90% to 100%</span>
     </p>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
