+++ public/test-coverage/Manager/dashboard.html
@@ -111,6 +111,7 @@
         </tr>
        </thead>
        <tbody>
+       <tr><td><a href="UserManager.php.html#50"><abbr title="App\Manager\UserManager::deleteUser">deleteUser</abbr></a></td><td class="text-right">0%</td></tr>
 
        </tbody>
       </table>
@@ -136,7 +137,7 @@
    <footer>
     <hr/>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
@@ -155,7 +156,7 @@ $(document).ready(function() {
       .yAxis.tickFormat(d3.format('d'));
 
     d3.select('#classCoverageDistribution svg')
-      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,0,0,2], "Class Coverage"))
+      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,0,1,1], "Class Coverage"))
       .transition().duration(500).call(chart);
 
     nv.utils.windowResize(chart.update);
@@ -173,7 +174,7 @@ $(document).ready(function() {
       .yAxis.tickFormat(d3.format('d'));
 
     d3.select('#methodCoverageDistribution svg')
-      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,0,0,10], "Method Coverage"))
+      .datum(getCoverageDistributionData([1,0,0,0,0,0,0,0,0,0,0,9], "Method Coverage"))
       .transition().duration(500).call(chart);
 
     nv.utils.windowResize(chart.update);
@@ -223,7 +224,7 @@ $(document).ready(function() {
     chart.yAxis.axisLabel('Cyclomatic Complexity');
 
     d3.select('#classComplexity svg')
-      .datum(getComplexityData([[100,8,"<a href=\"TaskManager.php.html#12\">App\\Manager\\TaskManager<\/a>"],[100,4,"<a href=\"UserManager.php.html#11\">App\\Manager\\UserManager<\/a>"]], 'Class Complexity'))
+      .datum(getComplexityData([[100,8,"<a href=\"TaskManager.php.html#12\">App\\Manager\\TaskManager<\/a>"],[94.44444444444444,4,"<a href=\"UserManager.php.html#11\">App\\Manager\\UserManager<\/a>"]], 'Class Complexity'))
       .transition()
       .duration(500)
       .call(chart);
@@ -247,7 +248,7 @@ $(document).ready(function() {
     chart.yAxis.axisLabel('Method Complexity');
 
     d3.select('#methodComplexity svg')
-      .datum(getComplexityData([[100,1,"<a href=\"TaskManager.php.html#14\">App\\Manager\\TaskManager::__construct<\/a>"],[100,1,"<a href=\"TaskManager.php.html#21\">App\\Manager\\TaskManager::createTask<\/a>"],[100,1,"<a href=\"TaskManager.php.html#29\">App\\Manager\\TaskManager::editTask<\/a>"],[100,1,"<a href=\"TaskManager.php.html#36\">App\\Manager\\TaskManager::deleteTask<\/a>"],[100,1,"<a href=\"TaskManager.php.html#41\">App\\Manager\\TaskManager::toggle<\/a>"],[100,3,"<a href=\"TaskManager.php.html#49\">App\\Manager\\TaskManager::convertStatusTaskToString<\/a>"],[100,1,"<a href=\"UserManager.php.html#13\">App\\Manager\\UserManager::__construct<\/a>"],[100,1,"<a href=\"UserManager.php.html#21\">App\\Manager\\UserManager::createUser<\/a>"],[100,1,"<a href=\"UserManager.php.html#34\">App\\Manager\\UserManager::editUser<\/a>"],[100,1,"<a href=\"UserManager.php.html#47\">App\\Manager\\UserManager::deleteUser<\/a>"]], 'Method Complexity'))
+      .datum(getComplexityData([[100,1,"<a href=\"TaskManager.php.html#17\">App\\Manager\\TaskManager::__construct<\/a>"],[100,1,"<a href=\"TaskManager.php.html#22\">App\\Manager\\TaskManager::createTask<\/a>"],[100,1,"<a href=\"TaskManager.php.html#30\">App\\Manager\\TaskManager::editTask<\/a>"],[100,1,"<a href=\"TaskManager.php.html#37\">App\\Manager\\TaskManager::deleteTask<\/a>"],[100,1,"<a href=\"TaskManager.php.html#42\">App\\Manager\\TaskManager::toggle<\/a>"],[100,3,"<a href=\"TaskManager.php.html#50\">App\\Manager\\TaskManager::convertStatusTaskToString<\/a>"],[100,1,"<a href=\"UserManager.php.html#17\">App\\Manager\\UserManager::__construct<\/a>"],[100,1,"<a href=\"UserManager.php.html#24\">App\\Manager\\UserManager::createUser<\/a>"],[100,1,"<a href=\"UserManager.php.html#37\">App\\Manager\\UserManager::editUser<\/a>"],[0,1,"<a href=\"UserManager.php.html#50\">App\\Manager\\UserManager::deleteUser<\/a>"]], 'Method Complexity'))
       .transition()
       .duration(500)
       .call(chart);
