+++ public/test-coverage/Manager/index.html
@@ -45,29 +45,29 @@
       <tr>
        <td class="success">Total</td>
        <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="96.88" aria-valuemin="0" aria-valuemax="100" style="width: 96.88%">
+           <span class="sr-only">96.88% covered (success)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">31&nbsp;/&nbsp;31</div></td>
+       <td class="success small"><div align="right">96.88%</div></td>
+       <td class="success small"><div align="right">31&nbsp;/&nbsp;32</div></td>
        <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="90.00" aria-valuemin="0" aria-valuemax="100" style="width: 90.00%">
+           <span class="sr-only">90.00% covered (success)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">10&nbsp;/&nbsp;10</div></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+       <td class="success small"><div align="right">90.00%</div></td>
+       <td class="success small"><div align="right">9&nbsp;/&nbsp;10</div></td>
+       <td class="danger big">       <div class="progress">
+         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="50.00" aria-valuemin="0" aria-valuemax="100" style="width: 50.00%">
+           <span class="sr-only">50.00% covered (danger)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">2&nbsp;/&nbsp;2</div></td>
+       <td class="danger small"><div align="right">50.00%</div></td>
+       <td class="danger small"><div align="right">1&nbsp;/&nbsp;2</div></td>
       </tr>
 
       <tr>
@@ -101,29 +101,29 @@
       <tr>
        <td class="success"><img src="../_icons/file-code.svg" class="octicon" /><a href="UserManager.php.html">UserManager.php</a></td>
        <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="94.44" aria-valuemin="0" aria-valuemax="100" style="width: 94.44%">
+           <span class="sr-only">94.44% covered (success)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">17&nbsp;/&nbsp;17</div></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+       <td class="success small"><div align="right">94.44%</div></td>
+       <td class="success small"><div align="right">17&nbsp;/&nbsp;18</div></td>
+       <td class="warning big">       <div class="progress">
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="75.00" aria-valuemin="0" aria-valuemax="100" style="width: 75.00%">
+           <span class="sr-only">75.00% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">4&nbsp;/&nbsp;4</div></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
+       <td class="warning small"><div align="right">75.00%</div></td>
+       <td class="warning small"><div align="right">3&nbsp;/&nbsp;4</div></td>
+       <td class="danger big">       <div class="progress">
+         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
+           <span class="sr-only">0.00% covered (danger)</span>
          </div>
        </div>
 </td>
-       <td class="success small"><div align="right">100.00%</div></td>
-       <td class="success small"><div align="right">1&nbsp;/&nbsp;1</div></td>
+       <td class="danger small"><div align="right">0.00%</div></td>
+       <td class="danger small"><div align="right">0&nbsp;/&nbsp;1</div></td>
       </tr>
 
 
@@ -139,7 +139,7 @@
      <span class="success"><strong>High</strong>: 90% to 100%</span>
     </p>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
