+++ public/test-coverage/Repository/dashboard.html
@@ -58,7 +58,7 @@
         </tr>
        </thead>
        <tbody>
-       <tr><td><a href="UserRepository.php.html#21">App\Repository\UserRepository</a></td><td class="text-right">85%</td></tr>
+       <tr><td><a href="UserRepository.php.html#21">App\Repository\UserRepository</a></td><td class="text-right">77%</td></tr>
        <tr><td><a href="TaskRepository.php.html#19">App\Repository\TaskRepository</a></td><td class="text-right">89%</td></tr>
 
        </tbody>
@@ -115,6 +115,7 @@
         </tr>
        </thead>
        <tbody>
+       <tr><td><a href="UserRepository.php.html#78"><abbr title="App\Repository\UserRepository::remove">remove</abbr></a></td><td class="text-right">0%</td></tr>
        <tr><td><a href="UserRepository.php.html#31"><abbr title="App\Repository\UserRepository::upgradePassword">upgradePassword</abbr></a></td><td class="text-right">80%</td></tr>
        <tr><td><a href="UserRepository.php.html#42"><abbr title="App\Repository\UserRepository::findByPagination">findByPagination</abbr></a></td><td class="text-right">81%</td></tr>
        <tr><td><a href="TaskRepository.php.html#26"><abbr title="App\Repository\TaskRepository::findByPagination">findByPagination</abbr></a></td><td class="text-right">86%</td></tr>
@@ -146,7 +147,7 @@
    <footer>
     <hr/>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
@@ -165,7 +166,7 @@ $(document).ready(function() {
       .yAxis.tickFormat(d3.format('d'));
 
     d3.select('#classCoverageDistribution svg')
-      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,2,0,0], "Class Coverage"))
+      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,1,1,0,0], "Class Coverage"))
       .transition().duration(500).call(chart);
 
     nv.utils.windowResize(chart.update);
@@ -183,7 +184,7 @@ $(document).ready(function() {
       .yAxis.tickFormat(d3.format('d'));
 
     d3.select('#methodCoverageDistribution svg')
-      .datum(getCoverageDistributionData([0,0,0,0,0,0,0,0,0,3,0,8], "Method Coverage"))
+      .datum(getCoverageDistributionData([1,0,0,0,0,0,0,0,0,3,0,7], "Method Coverage"))
       .transition().duration(500).call(chart);
 
     nv.utils.windowResize(chart.update);
@@ -233,7 +234,7 @@ $(document).ready(function() {
     chart.yAxis.axisLabel('Cyclomatic Complexity');
 
     d3.select('#classComplexity svg')
-      .datum(getComplexityData([[89.65517241379311,8,"<a href=\"TaskRepository.php.html#19\">App\\Repository\\TaskRepository<\/a>"],[85.18518518518519,8,"<a href=\"UserRepository.php.html#21\">App\\Repository\\UserRepository<\/a>"]], 'Class Complexity'))
+      .datum(getComplexityData([[89.65517241379311,8,"<a href=\"TaskRepository.php.html#19\">App\\Repository\\TaskRepository<\/a>"],[77.77777777777779,8,"<a href=\"UserRepository.php.html#21\">App\\Repository\\UserRepository<\/a>"]], 'Class Complexity'))
       .transition()
       .duration(500)
       .call(chart);
@@ -257,7 +258,7 @@ $(document).ready(function() {
     chart.yAxis.axisLabel('Method Complexity');
 
     d3.select('#methodComplexity svg')
-      .datum(getComplexityData([[100,1,"<a href=\"TaskRepository.php.html#21\">App\\Repository\\TaskRepository::__construct<\/a>"],[86.95652173913044,4,"<a href=\"TaskRepository.php.html#26\">App\\Repository\\TaskRepository::findByPagination<\/a>"],[100,1,"<a href=\"TaskRepository.php.html#61\">App\\Repository\\TaskRepository::add<\/a>"],[100,1,"<a href=\"TaskRepository.php.html#67\">App\\Repository\\TaskRepository::update<\/a>"],[100,1,"<a href=\"TaskRepository.php.html#72\">App\\Repository\\TaskRepository::remove<\/a>"],[100,1,"<a href=\"UserRepository.php.html#23\">App\\Repository\\UserRepository::__construct<\/a>"],[80,2,"<a href=\"UserRepository.php.html#31\">App\\Repository\\UserRepository::upgradePassword<\/a>"],[81.25,2,"<a href=\"UserRepository.php.html#42\">App\\Repository\\UserRepository::findByPagination<\/a>"],[100,1,"<a href=\"UserRepository.php.html#67\">App\\Repository\\UserRepository::add<\/a>"],[100,1,"<a href=\"UserRepository.php.html#73\">App\\Repository\\UserRepository::update<\/a>"],[100,1,"<a href=\"UserRepository.php.html#78\">App\\Repository\\UserRepository::remove<\/a>"]], 'Method Complexity'))
+      .datum(getComplexityData([[100,1,"<a href=\"TaskRepository.php.html#21\">App\\Repository\\TaskRepository::__construct<\/a>"],[86.95652173913044,4,"<a href=\"TaskRepository.php.html#26\">App\\Repository\\TaskRepository::findByPagination<\/a>"],[100,1,"<a href=\"TaskRepository.php.html#61\">App\\Repository\\TaskRepository::add<\/a>"],[100,1,"<a href=\"TaskRepository.php.html#67\">App\\Repository\\TaskRepository::update<\/a>"],[100,1,"<a href=\"TaskRepository.php.html#72\">App\\Repository\\TaskRepository::remove<\/a>"],[100,1,"<a href=\"UserRepository.php.html#23\">App\\Repository\\UserRepository::__construct<\/a>"],[80,2,"<a href=\"UserRepository.php.html#31\">App\\Repository\\UserRepository::upgradePassword<\/a>"],[81.25,2,"<a href=\"UserRepository.php.html#42\">App\\Repository\\UserRepository::findByPagination<\/a>"],[100,1,"<a href=\"UserRepository.php.html#67\">App\\Repository\\UserRepository::add<\/a>"],[100,1,"<a href=\"UserRepository.php.html#73\">App\\Repository\\UserRepository::update<\/a>"],[0,1,"<a href=\"UserRepository.php.html#78\">App\\Repository\\UserRepository::remove<\/a>"]], 'Method Complexity'))
       .transition()
       .duration(500)
       .call(chart);
