+++ public/test-coverage/Repository/index.html
@@ -45,21 +45,21 @@
       <tr>
        <td class="warning">Total</td>
        <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="87.50" aria-valuemin="0" aria-valuemax="100" style="width: 87.50%">
-           <span class="sr-only">87.50% covered (warning)</span>
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="83.93" aria-valuemin="0" aria-valuemax="100" style="width: 83.93%">
+           <span class="sr-only">83.93% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="warning small"><div align="right">87.50%</div></td>
-       <td class="warning small"><div align="right">49&nbsp;/&nbsp;56</div></td>
+       <td class="warning small"><div align="right">83.93%</div></td>
+       <td class="warning small"><div align="right">47&nbsp;/&nbsp;56</div></td>
        <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="72.73" aria-valuemin="0" aria-valuemax="100" style="width: 72.73%">
-           <span class="sr-only">72.73% covered (warning)</span>
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="63.64" aria-valuemin="0" aria-valuemax="100" style="width: 63.64%">
+           <span class="sr-only">63.64% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="warning small"><div align="right">72.73%</div></td>
-       <td class="warning small"><div align="right">8&nbsp;/&nbsp;11</div></td>
+       <td class="warning small"><div align="right">63.64%</div></td>
+       <td class="warning small"><div align="right">7&nbsp;/&nbsp;11</div></td>
        <td class="danger big">       <div class="progress">
          <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
            <span class="sr-only">0.00% covered (danger)</span>
@@ -101,21 +101,21 @@
       <tr>
        <td class="warning"><img src="../_icons/file-code.svg" class="octicon" /><a href="UserRepository.php.html">UserRepository.php</a></td>
        <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="85.19" aria-valuemin="0" aria-valuemax="100" style="width: 85.19%">
-           <span class="sr-only">85.19% covered (warning)</span>
+         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="77.78" aria-valuemin="0" aria-valuemax="100" style="width: 77.78%">
+           <span class="sr-only">77.78% covered (warning)</span>
          </div>
        </div>
 </td>
-       <td class="warning small"><div align="right">85.19%</div></td>
-       <td class="warning small"><div align="right">23&nbsp;/&nbsp;27</div></td>
-       <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="66.67" aria-valuemin="0" aria-valuemax="100" style="width: 66.67%">
-           <span class="sr-only">66.67% covered (warning)</span>
+       <td class="warning small"><div align="right">77.78%</div></td>
+       <td class="warning small"><div align="right">21&nbsp;/&nbsp;27</div></td>
+       <td class="danger big">       <div class="progress">
+         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="50.00" aria-valuemin="0" aria-valuemax="100" style="width: 50.00%">
+           <span class="sr-only">50.00% covered (danger)</span>
          </div>
        </div>
 </td>
-       <td class="warning small"><div align="right">66.67%</div></td>
-       <td class="warning small"><div align="right">4&nbsp;/&nbsp;6</div></td>
+       <td class="danger small"><div align="right">50.00%</div></td>
+       <td class="danger small"><div align="right">3&nbsp;/&nbsp;6</div></td>
        <td class="danger big">       <div class="progress">
          <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
            <span class="sr-only">0.00% covered (danger)</span>
@@ -139,7 +139,7 @@
      <span class="success"><strong>High</strong>: 90% to 100%</span>
     </p>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
