+++ public/test-coverage/Security/Voter/index.html
@@ -140,7 +140,7 @@
      <span class="success"><strong>High</strong>: 90% to 100%</span>
     </p>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 20:33:20 UTC 2024.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.31</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.8</a> and <a href="https://phpunit.de/">PHPUnit 9.6.19</a> at Tue Jun 11 23:08:10 UTC 2024.</small>
     </p>
    </footer>
   </div>
