+++ tests/Security/AccessTaskByUserTest.php
@@ -10,11 +10,13 @@ use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class AccessUserTest extends WebTestCase
+class AccessTaskByUserTest extends WebTestCase
 {
+
     private KernelBrowser $client;
     private TaskRepository $taskRepository;
     private UserRepository $userRepository;
+    
     private User $user;
     private User $userWithoutRole;
     private Task $taskUser1;
